pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Microstation"
version = "0.4.3"
authors = [{ name = "Dominik Reinartz", email = "dominik.reinartz@proton.me" }]
description = "Dynamically create Microcontroller layouts and connect them to your Computer."
readme = "README.md"
requires-python = ">=3.12"
keywords = ["microcontroller", "serial", "gui"]
license = { text = "MIT" }
dependencies = [
    "PyQt6>=6.8.0",
    "pyserial>=3.5",
    "pystray>=0.19.5",
    "pillow>=10.4.0",
    "pynput>=1.7.7",
    "psutil>=6.1.1",
    "pyudev>=0.24.3; sys_platform == 'linux'",
    "PyGObject>=3.50.0; sys_platform == 'linux'",
]

[project.optional-dependencies]
dev = [
    "qt6-tools>=6.5.0.1.3",
    "nuitka>=2.5.9",
    "flake8>=7.1.1",
    "mypy>=1.14.0",
    "isort>=5.13.2",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]

[project.scripts]
microstation = "microstation.__main__:main"

[tool.mypy]
files = ["microstation/"]
exclude = [
    "microstation/icons/",
    "microstation/ui/",
    "microstation/external_styles/",
]
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
follow_imports = "skip"
